+++ package.json
@@ -69,7 +69,7 @@
     "@swyg/corre": "^1.0.4",
     "@tanstack/react-query": "^5.64.1",
     "@untitled-ui/icons-react": "^0.1.1",
-    "@vitejs/plugin-react": "^4.3.2",
+    "@vitejs/plugin-react": "^4.3.4",
     "ao-tokens": "^0.0.4",
     "ar-gql": "1.2.9",
     "arbundles": "^0.9.5",
@@ -109,7 +109,7 @@
     "styled-components": "^5.3.6",
     "typed-assert": "^1.0.9",
     "uuid": "^9.0.0",
-    "vite-plugin-dts": "^4.5.0",
+    "vite-plugin-dts": "^4.5.3",
     "webextension-polyfill": "^0.10.0",
     "wouter": "^3.3.5"
   },
@@ -146,7 +146,8 @@
     "timers-browserify": "^2.0.12",
     "typescript": "^4.8.2",
     "url": "^0.11.0",
-    "vite": "^6.0.3",
+    "vite": "^6.2.5",
+    "vite-plugin-circular-dependency": "^0.5.0",
     "vite-plugin-node-stdlib-browser": "^0.2.1",
     "vitest": "^3.0.9",
     "vm-browserify": "^1.1.2"
