+++ src/components/dashboard/subsettings/AddToken.tsx
@@ -9,7 +9,11 @@ import {
 } from "@arconnect/components-rebrand";
 import browser from "webextension-polyfill";
 import { useEffect, useState } from "react";
-import { defaultTokens, type TokenInfo } from "~tokens/aoTokens/ao";
+import {
+  defaultTokens,
+  getTokenInfo,
+  type TokenInfo
+} from "~tokens/aoTokens/ao";
 import styled from "styled-components";
 import { isAddress } from "~utils/assertions";
 import { getAoTokens } from "~tokens";
@@ -19,7 +23,6 @@ import type { TokenType } from "~tokens/token";
 import { concatGatewayURL } from "~gateways/utils";
 import { FULL_HISTORY, useGateway } from "~gateways/wayfinder";
 import type { CommonRouteProps } from "~wallets/router/router.types";
-import { getTokenInfo } from "~tokens/aoTokens/router";
 import { AO_NATIVE_TOKEN } from "~utils/ao_import";
 import { ActionBar } from "~routes/popup/settings/tokens";
