+++ src/components/popup/home/AoBanner.tsx
@@ -4,9 +4,10 @@ import { Quantity } from "ao-tokens";
 import { useEffect, useState } from "react";
 import styled, { useTheme } from "styled-components";
 import browser from "webextension-polyfill";
-import { Id, useAo } from "~tokens/aoTokens/ao";
+import { Id } from "~tokens/aoTokens/ao";
 import { AO_NATIVE_TOKEN_BALANCE_MIRROR } from "~utils/ao_import";
 import { ExtensionStorage } from "~utils/storage";
+import { useAo } from "~tokens/hooks";
 
 interface AoBannerProps {
   activeAddress: string;
