+++ src/settings/hook.ts
@@ -1,6 +1,7 @@
 import { useStorage } from "~utils/storage";
 import { ExtensionStorage } from "~utils/storage";
-import { getSetting, PREFIX } from "~settings";
+import { getSetting } from "~settings";
+import { PREFIX } from "~settings/setting";
 
 const useSetting = <T = any>(name: string) => {
   const setting = getSetting(name);
