+++ src/utils/embedded/embedded.utils.ts
@@ -2,7 +2,12 @@ import { createSupabaseClient, createTRPCClient } from "embed-api";
 import { jwtDecode } from "jwt-decode";
 import { IS_EMBEDDED_APP } from "~utils/embedded/embedded.constants";
 import { LocalStorage } from "~iframe/storage/unpartitioned-storage/local-storage";
-import { searchParams, ancestorOrigin, isInsideIframe } from "./iframe.utils";
+import {
+  isInsideIframe,
+  EMBEDDED_ANCESTOR_ORIGIN,
+  EMBEDDED_CLIENT_ID,
+  EMBEDDED_SERVER_BASE_URL
+} from "./iframe.utils";
 import { ExtensionStorage } from "~utils/storage";
 import { postEmbeddedMessage } from "~utils/embedded/utils/messages/embedded-messages.utils";
 
@@ -13,60 +18,6 @@ import { postEmbeddedMessage } from "~utils/embedded/utils/messages/embedded-mes
 // - x-application-id (getApplicationIdHeader / setApplicationIdHeader)
 //
 // The code/functions below run in the context of Wander Embedded iframe/domain.
-
-const NODE_ENV = process.env.NODE_ENV || "development";
-
-const EMBEDDED_ENV_VARS_BY_ENV = {
-  development: {
-    DEFAULT_EMBEDDED_CLIENT_ID: import.meta.env
-      ?.VITE_DEV_DEFAULT_EMBEDDED_CLIENT_ID,
-    DEFAULT_EMBEDDED_SERVER_BASE_URL: import.meta.env
-      ?.VITE_DEV_DEFAULT_EMBEDDED_SERVER_BASE_URL
-  },
-  production: {
-    DEFAULT_EMBEDDED_CLIENT_ID: import.meta.env
-      ?.VITE_PROD_DEFAULT_EMBEDDED_CLIENT_ID,
-    DEFAULT_EMBEDDED_SERVER_BASE_URL: import.meta.env
-      ?.VITE_PROD_EMBEDDED_SERVER_BASE_URL
-  }
-} as const;
-
-const EMBEDDED_ENV_VARS = EMBEDDED_ENV_VARS_BY_ENV[NODE_ENV];
-
-if (!EMBEDDED_ENV_VARS)
-  throw new Error(`Missing ENV vars for NODE_ENV = "${NODE_ENV}"`);
-
-// Duplicated in `wander-embedded-sdk/src/utils/url/url.utils.ts`:
-const PARAM_CLIENT_ID = "client-id";
-const PARAM_SERVER_BASE_URL = "server-base-url";
-const PARAM_ANCESTOR_ORIGIN = "ancestor-origin";
-
-const EMBEDDED_CLIENT_ID =
-  searchParams.get(PARAM_CLIENT_ID) ||
-  EMBEDDED_ENV_VARS.DEFAULT_EMBEDDED_CLIENT_ID;
-
-const EMBEDDED_SERVER_BASE_URL =
-  searchParams.get(PARAM_SERVER_BASE_URL) ||
-  EMBEDDED_ENV_VARS.DEFAULT_EMBEDDED_SERVER_BASE_URL;
-
-const EMBEDDED_ANCESTOR_ORIGIN =
-  ancestorOrigin || searchParams.get(PARAM_ANCESTOR_ORIGIN);
-
-if (IS_EMBEDDED_APP) {
-  console.log("Wander Embedded URL params =", {
-    NODE_ENV,
-    EMBEDDED_CLIENT_ID,
-    EMBEDDED_SERVER_BASE_URL,
-    EMBEDDED_ANCESTOR_ORIGIN
-  });
-}
-
-// Note: DO NOT use document.referrer here as that will return the "incorrect" value when the user is redirected from
-// an auth provider domain to back to Wander Embedded.
-
-export function getEmbeddedAncestorOrigin() {
-  return EMBEDDED_ANCESTOR_ORIGIN;
-}
 
 // Note: This is run when trpc detects UNAUTHORIZED error.
 export async function signOut() {
