+++ src/utils/embedded/iframe.utils.ts
@@ -16,3 +16,57 @@ export function isInsideIframe(): boolean {
     return true;
   }
 }
+
+const NODE_ENV = process.env.NODE_ENV || "development";
+
+const EMBEDDED_ENV_VARS_BY_ENV = {
+  development: {
+    DEFAULT_EMBEDDED_CLIENT_ID: import.meta.env
+      ?.VITE_DEV_DEFAULT_EMBEDDED_CLIENT_ID,
+    DEFAULT_EMBEDDED_SERVER_BASE_URL: import.meta.env
+      ?.VITE_DEV_DEFAULT_EMBEDDED_SERVER_BASE_URL
+  },
+  production: {
+    DEFAULT_EMBEDDED_CLIENT_ID: import.meta.env
+      ?.VITE_PROD_DEFAULT_EMBEDDED_CLIENT_ID,
+    DEFAULT_EMBEDDED_SERVER_BASE_URL: import.meta.env
+      ?.VITE_PROD_EMBEDDED_SERVER_BASE_URL
+  }
+} as const;
+
+const EMBEDDED_ENV_VARS = EMBEDDED_ENV_VARS_BY_ENV[NODE_ENV];
+
+if (!EMBEDDED_ENV_VARS)
+  throw new Error(`Missing ENV vars for NODE_ENV = "${NODE_ENV}"`);
+
+// Duplicated in `wander-embedded-sdk/src/utils/url/url.utils.ts`:
+const PARAM_CLIENT_ID = "client-id";
+const PARAM_SERVER_BASE_URL = "server-base-url";
+const PARAM_ANCESTOR_ORIGIN = "ancestor-origin";
+
+export const EMBEDDED_CLIENT_ID =
+  searchParams.get(PARAM_CLIENT_ID) ||
+  EMBEDDED_ENV_VARS.DEFAULT_EMBEDDED_CLIENT_ID;
+
+export const EMBEDDED_SERVER_BASE_URL =
+  searchParams.get(PARAM_SERVER_BASE_URL) ||
+  EMBEDDED_ENV_VARS.DEFAULT_EMBEDDED_SERVER_BASE_URL;
+
+export const EMBEDDED_ANCESTOR_ORIGIN =
+  ancestorOrigin || searchParams.get(PARAM_ANCESTOR_ORIGIN);
+
+if (IS_EMBEDDED_APP) {
+  console.log("Wander Embedded URL params =", {
+    NODE_ENV,
+    EMBEDDED_CLIENT_ID,
+    EMBEDDED_SERVER_BASE_URL,
+    EMBEDDED_ANCESTOR_ORIGIN
+  });
+}
+
+// Note: DO NOT use document.referrer here as that will return the "incorrect" value when the user is redirected from
+// an auth provider domain to back to Wander Embedded.
+
+export function getEmbeddedAncestorOrigin() {
+  return EMBEDDED_ANCESTOR_ORIGIN;
+}
