+++ src/utils/embedded/utils/messages/embedded-messages.utils.ts
@@ -1,6 +1,8 @@
 import { nanoid } from "nanoid";
-import { getEmbeddedAncestorOrigin } from "~utils/embedded/embedded.utils";
-import { isInsideIframe } from "~utils/embedded/iframe.utils";
+import {
+  isInsideIframe,
+  getEmbeddedAncestorOrigin
+} from "~utils/embedded/iframe.utils";
 import type {
   EmbeddedCall,
   EmbeddedMessageId,
